+++ gps/msgfile/correlate.go
@@ -131,13 +131,14 @@ type Correlator struct {
 func NewCorrelator() *Correlator {
 	return &Correlator{
 		analyzers: map[gpsprot.Tag]responseAnalyzer{
-			gpsreg.TagUBX:          ubxAnalyzer{},
-			gpsreg.TagCASICBin:     casbinAnalyzer{},
-			gpsreg.TagAllystarBin:  asbinAnalyzer{},
-			gpsreg.TagSDBP:         sdbpAnalyzer{},
-			gpsreg.TagNMEA:         nmeaAnalyzer{},
-			gpsreg.TagUnicoreAscii: uncaAnalyzer{},
-			septentrio.TagReply:    septAnalyzer{},
+			gpsreg.TagUBX:                ubxAnalyzer{},
+			gpsreg.TagCASICBin:           casbinAnalyzer{},
+			gpsreg.TagAllystarBin:        asbinAnalyzer{},
+			gpsreg.TagSDBP:               sdbpAnalyzer{},
+			gpsreg.TagNMEA:               nmeaAnalyzer{},
+			gpsreg.TagUnicoreAscii:       uncaAnalyzer{},
+			gpsreg.TagNovAtelAbbrevAscii: novaaAnalyzer{},
+			septentrio.TagReply:          septAnalyzer{},
 		},
 	}
 }